Corrected optimization

+++ optimization/log_likelihood_grid/log_likelihood_grid.h
@@ -3,6 +3,8 @@
 
 #include <vector>
 
-std::vector<std::vector<double>> log_likelihood_grid(const std::vector<std::vector<double>>& xs, double threshold, const std::vector<double>& fixed_p2s);
+std::vector<std::vector<double> > log_likelihood_grid(const std::vector<std::vector<int> > &xs,
+                                                      double threshold,
+                                                      const std::vector<double> &fixed_p2s);
 
 #endif // LOG_LIKELIHOOD_GRID_H
