+++ optimization/maximize_product_grid/maximize_product_grid.h
@@ -2,8 +2,9 @@
 #define MAXIMIZE_PRODUCT_GRID_H
 
 #include <vector>
-#include <utility>
 
-std::vector<std::vector<double>> maximize_product_grid(const std::vector<std::vector<double>>& xs, double threshold, const std::vector<double>& fixed_p2s);
+std::vector<std::vector<double> > maximize_product_grid(const std::vector<std::vector<int> > &xs,
+                                                        double threshold,
+                                                        const std::vector<double> &fixed_p2s);
 
 #endif // MAXIMIZE_PRODUCT_GRID_H
